explicitly cast segment length from float to int

+++ Descriptors/JSONDescriptor.h
@@ -344,13 +344,13 @@ template <class DataFacadeT> class JSONDescriptor : public BaseDescriptor<DataFa
                     json_instruction_row.values.push_back(necessary_segments_running_index);
                     json_instruction_row.values.push_back(round(segment.duration / 10));
                     json_instruction_row.values.push_back(
-                        UintToString(static_cast<unsigned>(segment.length)) + "m");
+                        UintToString(static_cast<int>(segment.length)) + "m");
                     const double bearing_value = (segment.bearing / 10.) ;
                     json_instruction_row.values.push_back(Azimuth::Get(bearing_value));
                     json_instruction_row.values.push_back(static_cast<unsigned>(round(bearing_value)));
 
                     route_segments_list.emplace_back(
-                        segment.name_id, segment.length, route_segments_list.size());
+                        segment.name_id, static_cast<int>(segment.length), route_segments_list.size());
                     json_instruction_array.values.push_back(json_instruction_row);
                 }
             }
